+++ src/components/myPage/CurrentSubscription/index.js
@@ -1,0 +1,8 @@
+export { default as CurrentSubscription } from './Info';
+
+export { default as Fee } from './Fee';
+export { default as Unsubscriptions } from './Unsubscriptions';
+export { default as SubscriptionEditor } from './SubscriptionEditor';
+export { default as SubscriptionProviders } from './SubscriptionProviders';
+export { default as CurrentUnsubscriptionInfo } from './Unsubscriptions';
+export { default as ProviderBadges } from './ProviderBadges';
